Moved throwConfetti to utils
